Refactor event handler for userInput
The event handler for userInput was very messy and, as a result,
difficult to maintain. This proved especially tricky for updating the
value of num2 and displaying it in a clean way.

The refactored code handles number button clicks and operator button
clicks seperately. This helps in updating num2 for number button clicks
after an operator button has been clicked.