Fix bug when chaining different operations
There was a bug that would occur when chaining multiple operations with
different operators where 'result' wouldn't be the correct value. This
was due to the arguments for 'num1' and 'num2' that
'handleOperatoreInput' was using to call 'operate'. This fixes the bug
by checking the values of 'num1' and 'num2' before calling 'operate'.